에러 처리 로직 추가 및 관리자 권한 API 설정

+++ src/major-courses/major-courses.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { BadRequestException, ConflictException, Injectable } from '@nestjs/common';
 import { CreateMajorCourseDto } from './dto/create-major-course.dto';
 import { UpdateMajorCourseDto } from './dto/update-major-course.dto';
 import { InjectRepository } from '@nestjs/typeorm';
@@ -13,6 +13,16 @@ export class MajorCoursesService {
   ){}
 
   async createCourse(dto: CreateMajorCourseDto) {
+    const courseExists = await this.majorCourseRepository.exists({
+      where: {
+          course_id: dto.course_id,
+      }
+    });
+
+    if(courseExists){
+      throw new ConflictException('이미 존재하는 전공과목 입니다!')
+    }
+    
     const paper = this.majorCourseRepository.create(dto);
     return this.majorCourseRepository.save(paper);
   }
@@ -22,6 +32,16 @@ export class MajorCoursesService {
   }
 
   async updateCourse(course_id: string, dto: UpdateMajorCourseDto) {
+    const courseExists = await this.majorCourseRepository.exists({
+      where: {
+          course_id,
+      }
+    });
+
+    if(!courseExists){
+      throw new BadRequestException('존재하지 않는 전공과목 입니다!')
+    }
+
     await this.majorCourseRepository.update(course_id, dto);
     return this.majorCourseRepository.findOne({
       where: {
@@ -31,6 +51,18 @@ export class MajorCoursesService {
   }
 
   async deleteCourse(course_id: string) {
+    const courseExists = await this.majorCourseRepository.exists({
+      where: {
+          course_id,
+      }
+    });
+
+    if(!courseExists){
+      throw new BadRequestException('존재하지 않는 전공과목 입니다!')
+    }
+
     await this.majorCourseRepository.delete(course_id);
+
+    return course_id;
   }
 }
